requirements.txt
anyio==4.2.0
atomicwrites==1.4.1
attrs==23.2.0
Automat==20.2.0
beautifulsoup4==4.12.3
behave==1.2.6
bs4==0.0.2
certifi==2024.2.2
cffi==1.15.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==15.1.0
cryptography==36.0.2
cssselect==1.2.0
exceptiongroup==1.2.0
execnet==2.0.2
fasteners==0.19
filelock==3.13.1
flake8==4.0.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.2.0
incremental==21.3.0
iniconfig==2.0.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.8.0
markdown-it-py==3.0.0
mccabe==0.6.1
mdurl==0.1.2
outcome==1.3.0.post0
packaging==23.2
parameterized==0.9.0
parse==1.20.1
parse-type==0.6.2
parsel==1.6.0
pdbp==1.5.0
platformdirs==4.2.0
pluggy==1.4.0
priority==1.3.0
Protego==0.2.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
pyflakes==2.4.0
Pygments==2.17.2
pynose==1.4.8
pyOpenSSL==22.0.0
pyotp==2.9.0
pyparsing==3.0.7
pyreadline3==3.4.1
PySocks==1.7.1
pytest==8.0.0
pytest-html==2.0.1
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-pythonpath==0.7.4
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==20.8
PyYAML==6.0.1
queuelib==1.6.2
requests==2.31.0
rich==13.7.0
sbvirtualdisplay==1.3.0
Scrapy==2.5.1
selenium==4.17.2
seleniumbase==4.23.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tabcompleter==1.3.0
telegram==0.0.1
toml==0.10.2
tomli==2.0.1
transliterate==1.10.2
trio==0.24.0
trio-websocket==0.11.1
Twisted==22.2.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
urllib3==1.26.18
w3lib==1.22.0
webdriver-manager==4.0.1
wsproto==1.2.0
zipp==3.7.0
zope.interface==5.4.0
